Implemented logo mode. Obituaries are overlaid by game studios logos and disappear on hover.
