+++ ea-graveyard/src/studios.json
@@ -9,6 +9,7 @@
         "dateAcquired": "1998",
         "dateClosed": "2003-03-31",
         "link": "https://www.gamespot.com/articles/ea-consolidates-studios-closes-westwood/1100-2909852/",
+        "logo": "./public/logos/Westwood-Studios-1994.svg",
         "searchTags": [
             "Westwood Studios",
             "EA",
@@ -27,6 +28,7 @@
         "dateAcquired": "1997",
         "dateClosed": "2015-04-05",
         "link": "https://www.theverge.com/2015/3/4/8149827/ea-closing-maxis-emeryville",
+        "logo": "./public/logos/Maxis.svg",
         "searchTags": [
             "Maxis",
             "EA",
@@ -43,7 +45,8 @@
         "dateFounded": "1988",
         "dateAcquired": "2007",
         "dateClosed": "2009-11-17",
-        "link": "https://kotaku.com/confirmed-ea-closes-pandemic-studios-says-brand-will-5406830",
+        "link": "https://kotaku.com/confirmed-ea-closes-pandemic-studios-says-brand-will-5406830",  
+        "logo": "./public/logos/Pandemic-Studios.svg",
         "searchTags": [
             "Pandemic Studios",
             "EA",
@@ -62,6 +65,7 @@
         "dateAcquired": null,
         "dateClosed": "2017-10-17",
         "link": "https://kotaku.com/ea-shuts-down-visceral-games-1819623990",
+        "logo": "./public/logos/Visceral-Games.svg",
         "searchTags": [
             "Visceral Games",
             "EA",
@@ -82,6 +86,7 @@
         "dateAcquired": "2004-02-13",
         "dateClosed": "2007-11-06",
         "link": "https://www.gamespot.com/articles/ea-chicago-closing/1100-6182347/",
+        "logo": "./public/logos/NUFX.svg",
         "searchTags": [
             "NuFX",
             "EA Chicago",
@@ -102,6 +107,7 @@
         "dateAcquired": "2009-11-09",
         "dateClosed": "2013-06-14",
         "link": "https://www.gamesindustry.biz/ea-shutting-down-playfish-games",
+        "logo": "./public/logos/PlayFish.svg",
         "searchTags": [
             "Playfish",
             "EA",
@@ -119,6 +125,7 @@
         "dateAcquired": "2002-06-11",
         "dateClosed": "2013-04-13",
         "link": "https://www.polygon.com/2013/4/25/4266122/ea-layoffs-popcap-vancouver-quicklime",
+        "logo": "./public/logos/Black-Box.svg",
         "searchTags": [
             "Black Box Games",
             "EA",
@@ -135,6 +142,7 @@
         "dateAcquired": "2006-08-22",
         "dateClosed": "2013-07-12",
         "link": "https://www.gamesindustry.biz/ea-phenomic-closed",
+        "logo": "./public/logos/Phenomic.svg",
         "searchTags": [
             "Phenomic",
             "EA",
@@ -154,6 +162,7 @@
         "dateAcquired": "2000-02-24",
         "dateClosed": "2013-01-30",
         "link": "https://www.gamesindustry.biz/ea-phenomic-closed",
+        "logo": "./public/logos/DreamWorks-Interactive-1997.svg",
         "searchTags": [
             "DreamWorks Interactive",
             "EA",
@@ -173,6 +182,7 @@
         "dateAcquired": "1992-09-25",
         "dateClosed": "2004-03-01",
         "link": "https://www.gamesindustry.biz/ea-confirms-closure-of-origin-systems",
+        "logo": "./public/logos/Origin.svg",
         "searchTags": [
             "Origin Systems",
             "EA",
@@ -191,6 +201,7 @@
         "dateAcquired": "1995-01-06",
         "dateClosed": "2001",
         "link": "https://kotaku.com/one-explanation-behind-ea-destroying-bullfrog-1548092283",
+        "logo": "./public/logos/Bullfrog.svg",
         "searchTags": [
             "Bullfrog Productions",
             "EA",
@@ -209,6 +220,7 @@
         "dateAcquired": "1995-01-06",
         "dateClosed": "2001",
         "link": "https://www.gamesindustry.biz/mythic-shut-down",
+        "logo": "./public/logos/Mythic-Entertainment.svg",
         "searchTags": [
             "Mythic Entertainment",
             "EA",
